feat: encourage random ports via default
The idea is that this project provides a set of testing primitives (and
supporting helpers) - so we expect callers to be in a position to test
their subject with appropriate (test case specific) configuration. This
deters, but does not prevent, the use of static ports for some test
scenarios by way of included options (eg: for some external process with
limited reconfiguration support).

Signed-off-by: Jacob Vallejo <jake@jahkeup.com>